+++ src/player/Breeder.cpp
@@ -100,7 +100,7 @@ void Breeder::harvestLivestock() {
         int choice;
         cout << "Pilih Nomor hewan yang ingin dipanen: ";
         cin >> choice;
-        if (cin.fail() || choice < 1 || choice > harvestOptions.size()) {
+        if (cin.fail() || choice < 1 || choice > (int) harvestOptions.size()) {
             cin.clear(); 
             cin.ignore(numeric_limits<streamsize>::max(), '\n'); 
             throw invalid_argument("Input tidak valid. Harap masukkan nomor yang benar.");
